refactor: backend server와 연결하여 사용하도록 수정
